completed landing and FAQ page
